Agrego primera versión del mapa a tablero

+++ src/main/java/edu/fiuba/algo3/Jugador.java
@@ -44,6 +44,6 @@ public class Jugador {
         return this.vehiculo;
     }
     public void modificarPorcentajeMovimientos(float porcentaje) {
-        this.movimientos = (int) Math.round(this.movimientos * porcentaje);
+        this.movimientos = Math.round(this.movimientos * porcentaje);
     }
 }
